Add MainQueue and PreferenceQueue
