+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -150,11 +150,6 @@ public class AddCommandTest {
         }
 
         @Override
-        public boolean hasPatientInPatientQueue() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
         public ObservableList<Person> getFilteredPersonList() {
             throw new AssertionError("This method should not be called.");
         }
@@ -198,6 +193,11 @@ public class AddCommandTest {
 
         @Override
         public void commitAddressBook() {
+            throw new AssertionError("This method should not be called.");
+        }
+        //@@author iamjackslayer
+        @Override
+        public boolean hasPatientInMainQueue() {
             throw new AssertionError("This method should not be called.");
         }
     }
