+++ src/test/java/seedu/address/model/patientqueue/PreferenceQueueTest.java
@@ -1,0 +1,14 @@
+package seedu.address.model.patientqueue;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static seedu.address.testutil.TypicalPersons.ALICE;
+
+import org.junit.Test;
+
+import seedu.address.model.ModelManager;
+
+public class PreferenceQueueTest {
+    private ModelManager modelManager = new ModelManager();
+
+}
